+++ src/components/modal/modal.tsx
@@ -72,13 +72,7 @@ export const Modal: React.FC<ModalProps> = ({
     <div
       className={`${className} bg-white w-80 shadow-4 sans-serif relative`}
       data-testid="ipfs-modal"
-      style={{
-        maxWidth: '34em',
-        position: 'fixed',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%, -50%)'
-      }}
+      style={{ maxWidth: '34em' }}
       {...props}
     >
       {onCancel && (